move updateRegionCounters method to methods.js
