+++ collections/regions.js
@@ -1,5 +1,3 @@
-import '/imports/AsyncTools.js';
-
 // ======== DB-Model: ========
 // _id              -> ID
 // name             -> String
@@ -11,21 +9,3 @@ import '/imports/AsyncTools.js';
 
 Regions = new Meteor.Collection("Regions");
 if (Meteor.isServer) Regions._ensureIndex({loc : "2dsphere"});
-
-// We don't use AsyncTools.untilClean() here because consistency doesn't matter
-updateRegionCounters = function(regionId) {
-	var courseCount = Courses.find({ region: regionId }).count();
-	var futureEventCount = Events.find({ region: regionId, start: { $gte: new Date() } }).count();
-	Regions.update(regionId, { $set: {
-		courseCount: courseCount,
-		futureEventCount: futureEventCount
-	} });
-};
-
-Meteor.methods({
-	'updateRegionCounters': function(selector) {
-		Regions.find(selector).forEach(function(region) {
-			updateRegionCounters(region._id);
-		});
-	}
-});
