+++ server/start.js
@@ -83,7 +83,7 @@ Meteor.startup(function () {
 	// Update List of badges per user
 	Meteor.call('user.updateBadges', {}, AsyncTools.logErrors);
 
-	Meteor.call('updateRegionCounters', {}, AsyncTools.logErrors);
+	Meteor.call('regions.updateCounters', {}, AsyncTools.logErrors);
 
 	// Keep the nextEvent entry updated
 	// On startup do a full scan to catch stragglers
@@ -93,7 +93,7 @@ Meteor.startup(function () {
 			// Update nextEvent for courses where it expired
 			Meteor.call('updateNextEvent', { 'nextEvent.start': { $lt: new Date() }});
 
-			Meteor.call('updateRegionCounters', {}, AsyncTools.logErrors);
+			Meteor.call('regions.updateCounters', {}, AsyncTools.logErrors);
 		},
 		60*1000 // Check every minute
 	);
